requirements.txt
opencv-python~=4.8.1.78
numpy~=1.26.1
Pillow~=10.1.0
telebot~=0.0.5
config~=0.5.1
tensorflow~=2.14.0
handler~=1.3.0
keras~=2.14.0